requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
requests>=2.28
fpdf2>=2.7
PyPDF2>=3.0
Flask-WTF>=1.0
email-validator>=2.0
google-generativeai>=0.3.0
